ROSA provider: add a field to turn on/off vpc deletion
Previously when deleting a cluster and hosted control plane is enabled.
It will always try to delete the aws vpc that was created. Since the
create cluster allows the caller to supply subnet ids (by passing
creating a vpc), we need to be able to skip deleting vpc when subnet ids
are passed.

This commit introduces a new field to the delete cluster options struct.
Allowing the caller to turn off vpc deletion. Caller will know if they
provided subnet ids or had the vpc be created. Leaving the core method
to act on the input.