ROSA provider: add a field to turn on/off oidc config create/delete
By default if the caller provides no oidc config id, it will be created
and deleted. Some programs calling create/delete cluster methods, may
want to provide their own oidc config id to create the cluster. In those
cases, it may not want to be deleted afterwards "static oidc config".
This commit introduces a new field that the caller can set to disable
create/delete of the oidc config for the cluster.