ROSA provider: add option to use default account roles prefix
Previously when creating a cluster, it would create a new set of account
roles for each cluster, then delete them when the cluster is destroyed.

This commit adds a new field to the create cluster options struct,
allowing the caller to use the default account roles prefix
"ManagedOpenShift-X.Y" over creating unique ones per cluster. By
default, this is disabled and needs to be enabled.