pkg/openshift/rosa: new artifact dir field for create/delete clusters
Previously you were able to provide a working directory field when
creating/deleting clusters. This allowed the provider to write various
files (e.g. aws vpc terraform files, cluster install/uninstall logs) to
disk for later consumption.

There may be certain files the provider may not want to have archived and
just temporarily written to disk. This commit introduces a new field to
create/delete clusters options `ArtifactDir`. This directory is where
artifacts can be written/saved off for later usage. While the
`WorkingDir` is meant to be a temporary location and files are not
archived. With this change `WorkingDir` will write files such as
terraform files and `ArtifactDir` will store files such as
install/uninstall logs. When this field is not provided, it defaults to
the file systems temporary directory.