pkg/openshift/rosa: download rosa tar file to tmp directory
Previously the rosa tar file was being downloaded to the file path where
the go file is run. This fails when running within a prow CI job. To
workaround this, we can download the tar file to the tmp directory of
the container. Similar to where the actual rosa binary cli gets
installed to.